fix: Fix scroll to bottom does not transition state
This is due to BLoC thinking that it is emitting the same state, thus
does not re-emit the state with the new feed content. It is fixed by
adding updatedAt timestamp to the FeedLoadSuccess state, to trick BLoC
into recognising that it is indeed a new state, therefore emitting it.